+++ test/td_dq/download_test.exs
@@ -51,12 +51,18 @@ defmodule TdDq.DownloadTest do
     end
 
     test "to_csv/1 return csv content to download" do
-      content_labels = %{"quality_result.under_goal" => "Under Goal"}
-      header_labels = %{"template" => "Template Label"}
+      content_labels = %{
+        "quality_result.under_goal" => "Under Goal",
+        "executable.false" => "Internal",
+        "executable.true" => "Executable"
+      }
+
+      header_labels = %{"template" => "Template Label", "executable" => "Executable"}
 
       impl = %{
         implementation_key: "key1",
         implementation_type: "type1",
+        executable: true,
         rule: %{
           df_content: %{
             system: [
@@ -86,18 +92,24 @@ defmodule TdDq.DownloadTest do
 
       assert csv ==
                """
-               implementation_key;implementation_type;rule;Template Label;goal;minimum;business_concept;last_execution_at;result;execution;inserted_at;Info;System\r
-               #{impl.implementation_key};#{impl.implementation_type};#{impl.rule.name};#{impl.rule.df_name};#{impl.rule.goal};#{impl.rule.minimum};#{impl.current_business_concept_version.name};#{impl.execution_result_info.date};#{impl.execution_result_info.result};Under Goal;#{impl.inserted_at};field_value;system, system1\r
+               implementation_key;implementation_type;Executable;rule;Template Label;goal;minimum;business_concept;last_execution_at;result;execution;inserted_at;Info;System\r
+               #{impl.implementation_key};#{impl.implementation_type};Executable;#{impl.rule.name};#{impl.rule.df_name};#{impl.rule.goal};#{impl.rule.minimum};#{impl.current_business_concept_version.name};#{impl.execution_result_info.date};#{impl.execution_result_info.result};Under Goal;#{impl.inserted_at};field_value;system, system1\r
                """
     end
 
     test "to_csv/1 manages the download of uninformed result fields" do
-      content_labels = %{"quality_result.under_goal" => "Under Goal"}
-      header_labels = %{"template" => "Template Label"}
+      content_labels = %{
+        "quality_result.under_goal" => "Under Goal",
+        "executable.false" => "Internal",
+        "executable.true" => "Executable"
+      }
+
+      header_labels = %{"template" => "Template Label", "executable" => "Executable"}
 
       impl = %{
         implementation_key: "foo",
         implementation_type: "bar",
+        executable: true,
         rule: %{
           df_content: %{
             system: [
@@ -121,6 +133,7 @@ defmodule TdDq.DownloadTest do
       impl1 = %{
         implementation_key: "baz",
         implementation_type: "xyz",
+        executable: false,
         rule: %{
           df_content: %{
             system: [
@@ -148,9 +161,9 @@ defmodule TdDq.DownloadTest do
 
       assert csv ==
                """
-               implementation_key;implementation_type;rule;Template Label;goal;minimum;business_concept;last_execution_at;result;execution;inserted_at;Info;System\r
-               #{impl.implementation_key};#{impl.implementation_type};#{impl.rule.name};#{impl.rule.df_name};#{impl.rule.goal};#{impl.rule.minimum};#{impl.current_business_concept_version.name};;;;#{impl.inserted_at};field_value;system, system1\r
-               #{impl1.implementation_key};#{impl1.implementation_type};#{impl1.rule.name};#{impl1.rule.df_name};#{impl1.rule.goal};#{impl1.rule.minimum};#{impl1.current_business_concept_version.name};#{impl1.execution_result_info.date};#{impl1.execution_result_info.result};;#{impl1.inserted_at};field_value;system, system1\r
+               implementation_key;implementation_type;Executable;rule;Template Label;goal;minimum;business_concept;last_execution_at;result;execution;inserted_at;Info;System\r
+               #{impl.implementation_key};#{impl.implementation_type};Executable;#{impl.rule.name};#{impl.rule.df_name};#{impl.rule.goal};#{impl.rule.minimum};#{impl.current_business_concept_version.name};;;;#{impl.inserted_at};field_value;system, system1\r
+               #{impl1.implementation_key};#{impl1.implementation_type};Internal;#{impl1.rule.name};#{impl1.rule.df_name};#{impl1.rule.goal};#{impl1.rule.minimum};#{impl1.current_business_concept_version.name};#{impl1.execution_result_info.date};#{impl1.execution_result_info.result};;#{impl1.inserted_at};field_value;system, system1\r
                """
     end
   end