+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - [TD-3720] Update structure domain (with children)
 - [TD-3522] Support for StructureNote management with workflow
+- [TD-3552] Executable implementations
 
 ## [4.22.0] 2021-06-15